+++ src/i18n/ru.js
@@ -95,7 +95,7 @@ module.exports = ({ ADMIN }) => ({
   recent_acc:              'Acc',
   recent_combo:            'Комбо',
   recent_hits:             'Хиты',
-  recent_pp_unranked:      'карта не ранкед, в профиль не идёт',
+  pp_unranked_map:              'карта не ранкед, в профиль не идёт',
   recent_pp_failed:        'плей не доигран, значение только за сыгранную часть',
   recent_pp_failed_unranked: 'плей не доигран на карте не ранкед, только за сыгранную часть',
   recent_pass:             '✅ **Pass**',